% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/aster.R
\name{aster}
\alias{aster}
\title{aster HTMLwidget}
\usage{
aster(data = data.frame(id = c("FIS", "MAR", "AO", "NP", "CS"), order = c(1.1,
  1.3, 2, 3, 4), score = c(59, 24, 98, 60, 74), weight = c(0.5, 0.5, 1, 1, 1),
  color = c("#9E0041", "#C32F4B", "#E1514B", "#F47245", "#FB9F59"), label =
  c("Fisheries", "Mariculture", "Artisanal Fishing Opportunities",
  "Natural Products", "Carbon Storage")), width = NULL, height = NULL,
  background_color = "transparent", text_offset = 1, font_color = "black",
  tooltipStyle = "qtip-dark qtip-shadow qtip-tipsy", inner = 0.3,
  stroke = "grey", hover_color = "blue", font_size = "12px",
  font_size_center = "50px", margin_top = 40, margin_right = 80,
  margin_bottom = 40, margin_left = 80)
}
\arguments{
\item{data}{data frame holding chart data, expects columns id, order, score, weight, color and label}

\item{width}{chart width}

\item{height}{chart height}

\item{background_color}{chart background color, default transparent}

\item{text_offset}{value between 0 and +inf, values smaller than 1  place text labels inside slices, values larger than 1 place text labels outside slices, default 1}

\item{font_color}{label font color, default black}

\item{tooltipStyle}{tooltip style, see http://qtip2.com/options#style for details, defaults to 'qtip-dark qtip-shadow qtip-tipsy',}

\item{inner}{values smaller than 1 create donut charts (instead of pies), defaults to 0.3}

\item{stroke}{arc edge color, default grey}

\item{hover_color,}{slice hover color, default blue}

\item{font_size}{label text font size, default 12px}

\item{font_size_center}{font size numeric value in chart center, default 50px}

\item{margin_top}{chart top margin, default 40}

\item{margin_right}{chart right margin, default 80}

\item{margin_bottom}{chart bottom margin, default 40}

\item{margin_left}{chart left margin, default 80}
}
\description{
aster HTMLwidget
}

