% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/aster.R
\name{asterOutput}
\alias{asterOutput}
\title{Widget output function for use in Shiny}
\usage{
asterOutput(outputId, width = "500px", height = "500px")
}
\description{
Widget output function for use in Shiny
}

