% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/aster.R
\name{renderAster}
\alias{renderAster}
\title{Widget render function for use in Shiny}
\usage{
renderAster(expr, env = parent.frame(), quoted = FALSE)
}
\description{
Widget render function for use in Shiny
}

